Apply a hack solution to require_relative
Migrate to require_relative throughout the project.

For reasons unknown to man bash loses the call stack information when it returns
from the first require_relative call in a file and thus prevented the script from loading any further
relative libraries. The solution now keeps its own require_relative call stack and if
$(caller) and $BASH_SOURCE return empty or 'NULL', it will pop the call stack until
it finds the file where the require_relative call happened in its literal form. This
of course introduces some issues that need resolving, but it works as a proof of concept.